+++ db/dao/questions/questions_dao.go
@@ -6,7 +6,7 @@ import (
 )
 
 // CreateQuestion inserts a new question into the database.
-func CreateQuestion(questionSetID int, text string, timelimit int) (sql.Result, error) {
+func CreateQuestion(questionSetID int64, text string, timelimit int64) (sql.Result, error) {
 	db := db.GetDB()
 	res, err := db.Exec(
 		"INSERT INTO Questions (questionSetId, text, timelimit) VALUES ($1, $2, $3)",
@@ -17,7 +17,7 @@ func CreateQuestion(questionSetID int, text string, timelimit int) (sql.Result, 
 	return res, err
 }
 
-func GetQuestionByID(id int) (*sql.Row, error) {
+func GetQuestionByID(id int64) (*sql.Row, error) {
 	db := db.GetDB()
 	row := db.QueryRow("SELECT id, questionSetId, text, timelimit FROM Questions WHERE id = $1", id)
 	if row.Err() != nil {
@@ -26,7 +26,7 @@ func GetQuestionByID(id int) (*sql.Row, error) {
 	return row, nil
 }
 
-func GetQuestionsByQuestionSetID(questionSetID int) (*sql.Rows, error) {
+func GetQuestionsByQuestionSetID(questionSetID int64) (*sql.Rows, error) {
 	db := db.GetDB()
 	rows, err := db.Query("SELECT id, questionSetId, text, timelimit FROM Questions WHERE questionSetId = $1", questionSetID)
 	if err != nil {
@@ -36,7 +36,7 @@ func GetQuestionsByQuestionSetID(questionSetID int) (*sql.Rows, error) {
 }
 
 // UpdateQuestion updates an existing question in the database.
-func UpdateQuestion(id int, text string, timelimit int) error {
+func UpdateQuestion(id int64, text string, timelimit int64) error {
 	db := db.GetDB()
 	_, err := db.Exec("UPDATE Questions SET text = $1, timelimit = $2 WHERE id = $3", text, timelimit, id)
 	if err != nil {
@@ -47,7 +47,7 @@ func UpdateQuestion(id int, text string, timelimit int) error {
 }
 
 // DeleteQuestion marks a question as deleted in the database.
-func DeleteQuestion(id int) error {
+func DeleteQuestion(id int64) error {
 	db := db.GetDB()
 	// Prepare and execute the SQL UPDATE statement to set the "deleted" flag to true
 	_, err := db.Exec("DELETE FROM Questions WHERE id = $1", id)
